refactor: changed the middleware setup to be all in the middlewares directory
