--- middlewares/index.ts
@@ -1,5 +1,0 @@
-export { default as pluginDefault } from "./plugins/_pluginDefault";
-
-export { default as pluginContentSecurityPolicy } from "./plugins/pluginContentSecurityPolicy";
-export { default as pluginSupabase } from "./plugins/pluginSupabase";
-export { default as stack } from "./stack";
