+++ middlewares/plugins/index.ts
@@ -1,0 +1,4 @@
+export { default as pluginDefault } from "./_default";
+
+export { default as pluginContentSecurityPolicy } from "./pluginContentSecurityPolicy";
+export { default as pluginSupabase } from "./pluginSupabase";
